docs(query-keys): clarify query key refetching (#5653)
* docs(query-keys): clarify query key refetching 

Update the query-key documentation to note that changing data included in a query key will automatically refetch the query data.

It took me a long time to find this information, and I expected to find it here. I consulted with another engineer on our team, and they agreed that they would expect to see it stated more clearly here. So we felt this addition would be helpful.

* Update docs/react/guides/query-keys.md

* Update docs/react/guides/query-keys.md

---------

Co-authored-by: Dominik Dorfmeister <office@dorfmeister.cc>